pyproject.toml
[project]
name = "weav"
version = "0.1.0"
description = "Modern Python CLI application"
readme = "README.md"
authors = [{ name = "Henrik Holmboe", email = "henrik@dynamist.se" }]
requires-python = ">=3.11"
dependencies = [
  "typer>=0.15.0",
  "rich>=13.9.0",
  "jinja2>=3.1.0",
  "ruamel-yaml>=0.18.0",
  "platformdirs>=4.0.0",
]
license = { text = "Apache-2.0" }

[project.scripts]
weav = "weav.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "mypy>=1.13.0",
  "ruff>=0.8.0",
  "tox>=4.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "mypy>=1.13.0",
  "ruff>=0.8.0",
  "tox>=4.23.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["weav"]

[tool.ruff.lint]
select = [
  "ANN", # flake8-annotations
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "S",   # flake8-bandit (security)
  "C4",  # flake8-comprehensions
  "PTH", # flake8-use-pathlib
  "RUF", # ruff-specific rules
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "S101", # Allow assert statements in tests
  "ANN",  # Don't require type annotations in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
  "-ra",
  "-q",
  "--strict-markers",
  "--cov=weav",
  "--cov-report=term-missing",
  "--cov-report=html",
]

[tool.coverage.run]
source = ["weav"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if TYPE_CHECKING:",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]
